Updated the raw metadata code file so that the notebook with its outputs and evaluation results can be correctly rendered on GitHub interface.
The last commit was done after running all the cells in the finalzied Retriever Model notebook so that in the GitHub interface, the results of all the cells could be seen, including the reuslts of evaluation metrics. However, since the notebook used complex widgets (built-in) that could not be properly displayed on GitHub, we had to explicity remove the widgets metadata from the raw code file of the notebook.